+++ Test/Benchmark/MemoryTest.php
@@ -23,7 +23,7 @@ class MemoryTest extends \Magento\TestFramework\TestCase\AbstractController
     /** @var  ImporterFactory */
     private static $factory;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();