+++ Test/Integration/ReferenceResolverTest.php
@@ -16,7 +16,7 @@ class ReferenceResolverTest extends \Magento\TestFramework\TestCase\AbstractCont
     /** @var  ImporterFactory */
     private static $factory;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();