+++ Test/Integration/UrlKeyTest.php
@@ -34,7 +34,7 @@ class UrlKeyTest extends \Magento\TestFramework\TestCase\AbstractController
     /** @var  Metadata */
     protected static $metadata;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();