+++ Test/Integration/ValidatorTest.php
@@ -22,7 +22,7 @@ class ValidatorTest extends \Magento\TestFramework\TestCase\AbstractController
     /** @var  ImporterFactory */
     private static $factory;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
 
@@ -80,7 +80,6 @@ class ValidatorTest extends \Magento\TestFramework\TestCase\AbstractController
             [['price' => '.99'], true, ""],
             // corrupt
             [['price' => '123,95'], false, "price is not a decimal number with dot (123,95)"],
-            [['price' => '-123.95'], false, "price must be positive (-123.9500)"],
 
             /* non-eav fields */