+++ Test/Unit/ValueSerializerTest.php
@@ -10,7 +10,7 @@ use BigBridge\ProductImport\Model\Resource\Serialize\SerializeValueSerializer;
  */
 class ValueSerializerTest extends \PHPUnit\Framework\TestCase
 {
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         // include Magento auto-loading
         require_once __DIR__ . '/../../../../../app/autoload.php';